+++ 29-8 Ejercicio - Clase 3(Funciones)/matematicas.h
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void calcular(int num1, int num2, int operacion);
-float suma(int num1, int num2);
-float resta(int num1, int num2);
-float dividir(int num1, int num2);
-float multiplicar(int num1, int num2);
+void calcular(float num1, float num2, float operacion);
+float suma(float num1, float num2);
+float resta(float num1, float num2);
+float dividir(float num1, float num2);
+float multiplicar(float num1, float num2);
 void mostrar(float resultado);
 int pedirDato(int num1);
